+++ sei-db/bench/cryptosim/database.go
@@ -4,8 +4,8 @@ import (
 	"encoding/binary"
 	"fmt"
 
+	"github.com/sei-protocol/sei-chain/sei-db/bench/wrappers"
 	"github.com/sei-protocol/sei-chain/sei-db/proto"
-	"github.com/sei-protocol/sei-chain/sei-db/state_db/bench/wrappers"
 )
 
 // Encapsulates the database for the cryptosim benchmark.
