+++ sei-db/bench/cryptosim/historical_offload_test.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/sei-protocol/sei-chain/sei-db/state_db/bench/wrappers"
+	"github.com/sei-protocol/sei-chain/sei-db/bench/wrappers"
 )
 
 func TestValidateHistoricalOffloadRequiresConfigForHistoricalOffloadBackend(t *testing.T) {
